Quantai_Backend/backend/requirements.txt
Django>=5.2
djangorestframework>=3.14
django-oauth-toolkit>=1.7
django-user-agents
django-q2>=1.7.4
django-countries
whitenoise
geoip2
django-phone-field
requests
google-auth
google-auth-oauthlib
django-referrer-policy
PyJWT
social-auth-core
social-auth-app-django
cryptography
pillow
